removed settings loading responsabilities from build sources
